[Silver V] Title: 공포의 면담실, Time: 556 ms, Memory: 82036 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.java
@@ -21,8 +21,8 @@ public class Main {
 			}
 			queue.offer(sum);
 		}
-		int sum = 0;
-		int result = 0;
+		long sum = 0;
+		long result = 0;
 		while (!queue.isEmpty()) {
 			sum += queue.poll();
 			result += sum;
